pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nextcloud-mcp"
version = "0.1.0"
authors = [
  { name="Gemini Code Assist", email="author@example.com" },
]
description = "A Model-Context-Protocol (MCP) for saving files to Nextcloud and getting public share links."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = [
    "httpx[http2]>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
]
